chore: Update the version code to 11.1.0
### 重构功能
- 宝塔面板界面设置，支持主题导入导出

### 新增功能
- 暗色模式，支持多种颜色切换
- PHP项目-新增免费流量统计
- 节点管理-分发管理，支持编排命令和文件分发任务
- 专享版新增微信绑定
- 终端支持智能命令补全提示
- 网站-Nginx自定义Server段扩展脚本
- 证书验证方法兼容ecc私钥
- 网站-SSL页面新增https防窜站提示

### 优化功能
- 提升纯IPv6兼容性
- 系统防火墙端口转发规则的稳定性
- 大幅度PHP8.0+运行性能
- 深度对WordPress调优，大幅增强访问速度(需重装PHP8.0+)
- 节点管理-提升负载均衡稳定性
- 提升Docker模块的稳定性
- 设置-备份还原与整机迁移的稳定性
- 面板自动安全备份取消备份入侵检测日志，避免备份文过大
- 安全-SSH登录告警支持IP网段

### 修复问题
- 修改时区后面板不同步的问题
- ssl域名管理-解析记录管理表格分页问题
- 修复部分情况下添加域名提示错误的问题
- 修复部分情况下pip源错误导致的OpenSSL库无法修复的问题
- 其他已知BUG

+++ class/firewalls.py
@@ -7,11 +7,12 @@
 # | Author: hwliang <hwl@bt.cn>
 # +-------------------------------------------------------------------
 import sys,os,public,re,firewalld,time
+
 class firewalls:
     __isFirewalld = False
     __isUfw = False
     __Obj = None
-    
+
     def __init__(self):
         if os.path.exists('/usr/sbin/firewalld'): self.__isFirewalld = True
         if os.path.exists('/usr/sbin/ufw'): self.__isUfw = True
@@ -21,8 +22,8 @@ class firewalls:
                 self.GetList()
             except:
                 pass
-        
-    
+
+
     #获取服务端列表
     def GetList(self):
         try:
@@ -32,7 +33,7 @@ class firewalls:
             for i in range(len(data['ports'])):
                 tmp = self.CheckDbExists(data['ports'][i]['port'])
                 if not tmp: public.M('firewall').add('port,ps,addtime',(data['ports'][i]['port'],'',addtime))
-                          
+
             data['iplist'] = self.__Obj.GetDropAddressList()
             for i in range(len(data['iplist'])):
                 try:
@@ -42,41 +43,42 @@ class firewalls:
                     pass
         except:
             pass
-    
+
     #检查数据库是否存在
     def CheckDbExists(self,port):
         data = public.M('firewall').field('id,port,ps,addtime').select()
         for dt in data:
             if dt['port'] == port: return dt
         return False
-        
+
     #重载防火墙配置
     def FirewallReload(self):
         if self.__isUfw:
-            public.ExecShell('/usr/sbin/ufw reload')
+            public.ExecShell('/usr/sbin/ufw reload &')
             return
         if self.__isFirewalld:
-            public.ExecShell('firewall-cmd --reload')
+            public.ExecShell('firewall-cmd --reload &')
         else:
-            public.ExecShell('/etc/init.d/iptables save')
-            public.ExecShell('/etc/init.d/iptables restart')
+            public.ExecShell('/etc/init.d/iptables save &')
+            public.ExecShell('/etc/init.d/iptables restart &')
 
     #取防火墙状态
     def CheckFirewallStatus(self):
-        if self.__isUfw:
-            res = public.ExecShell('ufw status verbose')[0]
-            if res.find('inactive') != -1: return False
-            return True
+        # if self.__isUfw:
+        #     res = public.ExecShell('ufw status verbose')[0]
+        #     if res.find('inactive') != -1: return False
+        #     return True
 
-        if self.__isFirewalld:
-            res = public.ExecShell("systemctl status firewalld")[0]
-            if res.find('active (running)') != -1: return True
-            if res.find('disabled') != -1: return False
-            if res.find('inactive (dead)') != -1: return False
-        else:
-            res = public.ExecShell("/etc/init.d/iptables status")[0]
-            if res.find('not running') != -1: return False
-            return True
+        # if self.__isFirewalld:
+        #     res = public.ExecShell("systemctl status firewalld")[0]
+        #     if res.find('active (running)') != -1: return True
+        #     if res.find('disabled') != -1: return False
+        #     if res.find('inactive (dead)') != -1: return False
+        # else:
+        #     res = public.ExecShell("/etc/init.d/iptables status")[0]
+        #     if res.find('not running') != -1: return False
+        #     return True
+        return public.get_firewall_status() == 1
 
     def SetFirewallStatus(self,get=None):
         '''
@@ -106,7 +108,7 @@ class firewalls:
                 public.ExecShell('/etc/init.d/iptables stop')
         public.WriteLog('TYPE_FIREWALL','{}系统防火墙'.format(status_msg[status]))
         return public.returnMsg(True,'已{}系统防火墙'.format(status_msg[status]))
-        
+
     #添加屏蔽IP
     def AddDropAddress(self,get):
         if not self.CheckFirewallStatus(): return public.returnMsg(False,'当前系统防火墙未开启')
@@ -114,11 +116,15 @@ class firewalls:
         import re
         ip_format = get.port.split('/')[0]
         if not public.check_ip(ip_format): return public.returnMsg(False,'FIREWALL_IP_FORMAT')
+        # if public.is_ipv6(ip_format): return public.returnMsg(False,'暂不支持屏蔽IPv6')
         if ip_format in  ['0.0.0.0','127.0.0.0',"::1"]: return public.returnMsg(False,'请不要花样作死!')
         address = get.port
         if public.M('firewall').where("port=?",(address,)).count() > 0: return public.returnMsg(False,'FIREWALL_IP_EXISTS')
         if self.__isUfw:
-            public.ExecShell('ufw insert 1 deny from ' + address + ' to any')
+            if public.is_ipv6(ip_format):
+                public.ExecShell('ufw deny from ' + address + ' to any')
+            else:
+                public.ExecShell('ufw insert 1 deny from ' + address + ' to any')
         else:
             if self.__isFirewalld:
                 #self.__Obj.AddDropAddress(address)
@@ -129,39 +135,40 @@ class firewalls:
             else:
                 if public.is_ipv6(ip_format): return public.returnMsg(False,'FIREWALL_IP_FORMAT')
                 public.ExecShell('iptables -I INPUT -s '+address+' -j DROP')
-        
+
         public.WriteLog("TYPE_FIREWALL", 'FIREWALL_DROP_IP',(address,))
         addtime = time.strftime('%Y-%m-%d %X',time.localtime())
         public.M('firewall').add('port,ps,addtime',(address,get.ps,addtime))
         self.FirewallReload()
         return public.returnMsg(True,'ADD_SUCCESS')
-    
-    
+
+
     #删除IP屏蔽
     def DelDropAddress(self,get):
         if not self.CheckFirewallStatus(): return public.returnMsg(False,'当前系统防火墙未开启')
         address = get.port
+        
         id = get.id
         ip_format = get.port.split('/')[0]
+        if not public.check_ip(ip_format): return public.returnMsg(False,'FIREWALL_IP_FORMAT')
         if self.__isUfw:
             public.ExecShell('ufw delete deny from ' + address + ' to any')
         else:
             if self.__isFirewalld:
-                #self.__Obj.DelDropAddress(address)
                 if public.is_ipv6(ip_format):
                     public.ExecShell('firewall-cmd --permanent --remove-rich-rule=\'rule family=ipv6 source address="'+ address +'" drop\'')
                 else:
                     public.ExecShell('firewall-cmd --permanent --remove-rich-rule=\'rule family=ipv4 source address="'+ address +'" drop\'')
             else:
                 public.ExecShell('iptables -D INPUT -s '+address+' -j DROP')
-        
+
         public.WriteLog("TYPE_FIREWALL",'FIREWALL_ACCEPT_IP',(address,))
         public.M('firewall').where("id=?",(id,)).delete()
-        
+
         self.FirewallReload()
         return public.returnMsg(True,'DEL_SUCCESS')
-    
-    
+
+
     #添加放行端口
     def AddAcceptPort(self,get):
         if not self.CheckFirewallStatus(): return public.returnMsg(False,'当前系统防火墙未开启')
@@ -180,16 +187,16 @@ class firewalls:
         notudps = ['80','443','8888','888','39000:40000','21','22']
         if self.__isUfw:
             public.ExecShell('ufw allow ' + port + '/tcp')
-            if not port in notudps: public.ExecShell('ufw allow ' + port + '/udp')
+            # if not port in notudps: public.ExecShell('ufw allow ' + port + '/udp')
         else:
             if self.__isFirewalld:
                 #self.__Obj.AddAcceptPort(port)
                 port = port.replace(':','-')
                 public.ExecShell('firewall-cmd --permanent --zone=public --add-port='+port+'/tcp')
-                if not port in notudps: public.ExecShell('firewall-cmd --permanent --zone=public --add-port='+port+'/udp')
+                # if not port in notudps: public.ExecShell('firewall-cmd --permanent --zone=public --add-port='+port+'/udp')
             else:
                 public.ExecShell('iptables -I INPUT -p tcp -m state --state NEW -m tcp --dport '+port+' -j ACCEPT')
-                if not port in notudps: public.ExecShell('iptables -I INPUT -p tcp -m state --state NEW -m udp --dport '+port+' -j ACCEPT')
+                # if not port in notudps: public.ExecShell('iptables -I INPUT -p tcp -m state --state NEW -m udp --dport '+port+' -j ACCEPT')
         public.WriteLog("TYPE_FIREWALL", 'FIREWALL_ACCEPT_PORT',(port,))
         addtime = time.strftime('%Y-%m-%d %X',time.localtime())
         if not is_exists: public.M('firewall').add('port,ps,addtime',(port,ps,addtime))
@@ -207,22 +214,28 @@ class firewalls:
             return False
         if self.__isUfw:
             public.ExecShell('ufw allow ' + port + '/tcp')
-            public.ExecShell('ufw allow ' + port + '/udp')
+            # public.ExecShell('ufw allow ' + port + '/udp')
         else:
             if self.__isFirewalld:
                 port = port.replace(':','-')
                 public.ExecShell('firewall-cmd --permanent --zone=public --add-port='+port+'/tcp')
-                public.ExecShell('firewall-cmd --permanent --zone=public --add-port='+port+'/udp')
+                # public.ExecShell('firewall-cmd --permanent --zone=public --add-port='+port+'/udp')
             else:
                 public.ExecShell('iptables -I INPUT -p tcp -m state --state NEW -m tcp --dport '+port+' -j ACCEPT')
-                public.ExecShell('iptables -I INPUT -p tcp -m state --state NEW -m udp --dport '+port+' -j ACCEPT')
+                # public.ExecShell('iptables -I INPUT -p tcp -m state --state NEW -m udp --dport '+port+' -j ACCEPT')
         return True
-    
+
     #删除放行端口
     def DelAcceptPort(self,get):
         if not self.CheckFirewallStatus(): return public.returnMsg(False,'当前系统防火墙未开启')
         port = get.port
         id = get.id
+        port2 = port.replace('-',':')
+        rep = r"^\d{1,5}(:\d{1,5})?$"
+        if not re.search(rep,port2):
+            return public.returnMsg(False,'PORT_CHECK_RANGE')
+        if public.is_ipv6(port): return self.DelDropAddress(get) # 如果是ipv6地址，则调用DelDropAddress
+
         try:
             if(port == public.GetHost(True) or port == public.readFile('data/port.pl').strip()): return public.returnMsg(False,'FIREWALL_PORT_PANEL')
             if self.__isUfw:
@@ -238,56 +251,60 @@ class firewalls:
                     public.ExecShell('iptables -D INPUT -p tcp -m state --state NEW -m udp --dport '+port+' -j ACCEPT')
             public.WriteLog("TYPE_FIREWALL", 'FIREWALL_DROP_PORT',(port,))
             public.M('firewall').where("id=?",(id,)).delete()
-            
+
             self.FirewallReload()
             return public.returnMsg(True,'DEL_SUCCESS')
         except:
             return public.returnMsg(False,'DEL_ERROR')
-    
+
     #设置远程端口状态
     def SetSshStatus(self,get):
-        version = public.readFile('/etc/redhat-release')
+        # version = public.readFile('/etc/redhat-release')
         if int(get['status'])==1:
             msg = public.getMsg('FIREWALL_SSH_STOP')
             act = 'stop'
         else:
             msg = public.getMsg('FIREWALL_SSH_START')
             act = 'start'
-        
-        if not os.path.exists('/etc/redhat-release'):
-            public.ExecShell('service ssh ' + act)
-        elif version.find(' 7.') != -1 or version.find(' 8.') != -1 or version.find('Fedora') != -1:
-            public.ExecShell("systemctl "+act+" sshd.service")
-        else:
-            public.ExecShell("/etc/init.d/sshd "+act)
-        
+
+        public.set_sshd_status(act)
+        if act in ['start'] and not public.get_sshd_status():
+            msg = 'SSHD服务启动失败'
+            public.WriteLog("TYPE_FIREWALL", msg)
+            return public.returnMsg(False,msg)
         public.WriteLog("TYPE_FIREWALL", msg)
         return public.returnMsg(True,'SUCCESS')
 
-        
-    
-    
+
     #设置ping
-    def SetPing(self,get):
+    def SetPing(self, get):
+        if "status" not in get:
+            return public.returnMsg(False,'请传入status参数')
         if get.status == '1':
             get.status = '0'
         else:
             get.status = '1'
         filename = '/etc/sysctl.conf'
         conf = public.readFile(filename)
+        if not conf: return public.returnMsg(False, '/etc/sysctl.conf文件读取失败,无法设置禁ping!')
         if conf.find('net.ipv4.icmp_echo') != -1:
             rep = r"net\.ipv4\.icmp_echo.*"
-            conf = re.sub(rep,'net.ipv4.icmp_echo_ignore_all='+get.status,conf)
+            conf = re.sub(rep, 'net.ipv4.icmp_echo_ignore_all=' + get.status + "\n", conf)
         else:
-            conf += "\nnet.ipv4.icmp_echo_ignore_all="+get.status
-            
-        
-        public.writeFile(filename,conf)
-        public.ExecShell('sysctl -p')
-        return public.returnMsg(True,'SUCCESS') 
-        
-    
-    
+            conf += "\nnet.ipv4.icmp_echo_ignore_all=" + get.status + "\n"
+
+        if public.writeFile(filename, conf):
+            public.ExecShell('sysctl -p')
+            cache_key = "firewall_info"
+            from BTPanel import cache
+            data = cache.get(cache_key)
+            if data: cache.delete(cache_key)
+            return public.returnMsg(True, 'SUCCESS')
+        else:
+            return public.returnMsg(False,'<a style="color:red;">错误：设置失败，sysctl.conf不可写!</a><br>1、如果安装了[宝塔系统加固]，请先关闭<br>2、如果安装了云锁，请关闭[系统加固]功能<br>3、如果安装了安全狗，请关闭[系统防护]功能<br>4、如果使用了其它安全软件，请先卸载<br>')
+
+
+
     #改远程端口
     def SetSshPort(self,get):
         port = get.port
@@ -296,57 +313,51 @@ class firewalls:
         if port in ports: return public.returnMsg(False,'请不要使用常用程序的默认端口!')
         file = '/etc/ssh/sshd_config'
         conf = public.readFile(file)
-        
+        if not conf: return public.returnMsg(False, 'SSH配置文件异常，请手动重新安装SSH服务后再试!')
+
         rep = r"#*Port\s+([0-9]+)\s*\n"
         conf = re.sub(rep, "Port "+port+"\n", conf)
         public.writeFile(file,conf)
-        
+
+        #ubuntu 24版本起 还需要额外修改socket文件
+        osinfo = public.get_os_version().lower()
+        osinfo = osinfo.split(' ')
+        try:
+            if osinfo[0] == 'ubuntu' and osinfo[1].split('.')[0] > '22':
+                socket_file = '/lib/systemd/system/ssh.socket'
+                conf = public.readFile(socket_file)
+                if conf:
+                    rep = r"ListenStream=+([0-9]+)\s*\n"
+                    conf = re.sub(rep, "ListenStream="+port+"\n", conf)
+                    public.writeFile(socket_file, conf)
+                    public.ExecShell("systemctl daemon-reload;systemctl restart ssh.socket")
+        except:
+            pass
+
         if self.__isFirewalld:
             public.ExecShell('firewall-cmd --permanent --zone=public --add-port='+port+'/tcp')
             public.ExecShell('setenforce 0')
             public.ExecShell('sed -i "s#SELINUX=enforcing#SELINUX=disabled#" /etc/selinux/config')
-            public.ExecShell("systemctl restart sshd.service")
+            # public.ExecShell("systemctl restart sshd.service")
         elif self.__isUfw:
             public.ExecShell('ufw allow ' + port + '/tcp')
-            public.ExecShell("service ssh restart")
+            # public.ExecShell("service ssh restart")
         else:
             public.ExecShell('iptables -I INPUT -p tcp -m state --state NEW -m tcp --dport '+port+' -j ACCEPT')
-            public.ExecShell("/etc/init.d/sshd restart")
-        
+            # public.ExecShell("/etc/init.d/sshd restart")
+
+        public.set_sshd_status(status_act="restart")
+
         self.FirewallReload()
         public.M('firewall').where("ps=? or ps=? or port=?",('SSH远程管理服务','SSH远程服务',port)).delete()
         public.M('firewall').add('port,ps,addtime',(port,'SSH远程服务',time.strftime('%Y-%m-%d %X',time.localtime())))
         public.WriteLog("TYPE_FIREWALL", "FIREWALL_SSH_PORT",(port,))
-        return public.returnMsg(True,'EDIT_SUCCESS') 
-    
+        return public.returnMsg(True,'EDIT_SUCCESS')
+
     #取SSH信息
     def GetSshInfo(self,get):
-        port = public.get_ssh_port()
-
-        pid_file = '/run/sshd.pid'
-        if os.path.exists(pid_file):
-            pid = int(public.readFile(pid_file))
-            status = public.pid_exists(pid)
-        else:
-            import system
-            panelsys = system.system()
-            
-            version = panelsys.GetSystemVersion()
-            if os.path.exists('/usr/bin/apt-get'):
-                if os.path.exists('/etc/init.d/sshd'):
-                    status = public.ExecShell("service sshd status | grep -P '(dead|stop)'|grep -v grep")
-                else:
-                    status = public.ExecShell("service ssh status | grep -P '(dead|stop)'|grep -v grep")
-            else:
-                if version.find(' 7.') != -1 or version.find(' 8.') != -1 or version.find('Fedora') != -1:
-                    status = public.ExecShell("systemctl status sshd.service | grep 'dead'|grep -v grep")
-                else:
-                    status = public.ExecShell("/etc/init.d/sshd status | grep -e 'stopped' -e '已停'|grep -v grep")
-            
-            if len(status[0]) > 3:
-                status = False
-            else:
-                status = True
+        port = public.get_sshd_port()
+        status = public.get_sshd_status()
         isPing = True
         try:
             file = '/etc/sysctl.conf'
@@ -356,12 +367,15 @@ class firewalls:
             if tmp == '1': isPing = False
         except:
             isPing = True
-        
+
         data = {}
         data['port'] = port
         data['status'] = status
         data['ping'] = isPing
         data['firewall_status'] = self.CheckFirewallStatus()
         return data
-        
-    
+
+
+
+
+
